Split into lib and main
